+++ .gitattributes
@@ -19,6 +19,6 @@
 /API.md linguist-generated
 /LICENSE linguist-generated
 /package.json linguist-generated
-/src\lambda\custom-resource-function.ts linguist-generated
+/src\custom-resource-function.ts linguist-generated
 /tsconfig.dev.json linguist-generated
 /yarn.lock linguist-generated
