+++ .gitignore
@@ -53,4 +53,4 @@ junit.xml
 tsconfig.json
 !/API.md
 /assets/
-!/src\lambda\custom-resource-function.ts
+!/src\custom-resource-function.ts
