+++ .prettierignore
@@ -1,2 +1,2 @@
 # ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
-src\lambda\custom-resource-function.ts
+src\custom-resource-function.ts
