+++ package.json
@@ -8,8 +8,8 @@
     "build": "npx projen build",
     "bump": "npx projen bump",
     "bundle": "npx projen bundle",
-    "bundle:lambda\\custom-resource.lambda": "npx projen bundle:lambda\\custom-resource.lambda",
-    "bundle:lambda\\custom-resource.lambda:watch": "npx projen bundle:lambda\\custom-resource.lambda:watch",
+    "bundle:custom-resource.lambda": "npx projen bundle:custom-resource.lambda",
+    "bundle:custom-resource.lambda:watch": "npx projen bundle:custom-resource.lambda:watch",
     "clobber": "npx projen clobber",
     "compat": "npx projen compat",
     "compile": "npx projen compile",
